addition of helper modules. Modifications to extractor classes
