+++ lib/cuke_slicer.rb
@@ -2,10 +2,19 @@ require 'cuke_modeler'
 
 require "cuke_slicer/version"
 require "cuke_slicer/slicer"
+
+
+require "cuke_slicer/helpers/helpers"
+require "cuke_slicer/helpers/extraction_helpers"
+require "cuke_slicer/helpers/filter_helpers"
+require "cuke_slicer/helpers/matching_helpers"
+
 require "cuke_slicer/filter_set"
+
 require "cuke_slicer/collections/tag_collection"
 require "cuke_slicer/collections/nested_tag_collection"
 require "cuke_slicer/collections/path_collection"
+
 require "cuke_slicer/extractors/directory_extractor"
 require "cuke_slicer/extractors/file_extractor"
