+++ lib/cuke_slicer/extractors/directory_extractor.rb
@@ -25,6 +25,7 @@ module CukeSlicer
             else
               # Non-feature files are ignored
           end
+
         end
       end
     end
