+++ lib/cuke_slicer/filter_set.rb
@@ -1,5 +1,6 @@
 module CukeSlicer
   class FilterSet
+    include Helpers
     attr_accessor :filter_type, :filter_value
 
     def initialize filter_type, filter_value
@@ -27,22 +28,6 @@ module CukeSlicer
 
     def block_invalid
       raise(ArgumentError, "Invalid filter '#{filter_value}'. Must be a String, Regexp, or Array thereof. Got #{filter_value.class}") unless str_regex_arr?(filter_value)
-    end
-
-    def str_regex?(parameter)
-      parameter.is_a?(String) or parameter.is_a?(Regexp)
-    end
-
-    def str_regex_arr?(parameter)
-      parameter.is_a?(String) or parameter.is_a?(Regexp) or parameter.is_a?(Array)
-    end
-
-    def is_tag?(parameter)
-      parameter.to_s =~ /tag/
-    end
-
-    def is_path?(parameter)
-      parameter.to_s =~ /path/
     end
 
   end
