+++ lib/cuke_slicer/helpers/matching_helpers.rb
@@ -1,0 +1,45 @@
+module CukeSlicer
+  module MatchingHelpers
+    def matching_tag?(element, filters)
+      filters.each do |filter|
+        if filter.is_a?(Array)
+          filter_match = or_filter_match(element, filter)
+        else
+          filter_match = and_filter_match(element, filter)
+        end
+
+        return false unless filter_match
+      end
+
+      true
+    end
+
+    def and_filter_match(element, filter)
+      filter_match(element, filter)
+    end
+
+    def or_filter_match(element, filters)
+      filters.any? do |filter|
+        filter_match(element, filter)
+      end
+    end
+
+    def filter_match(element, filter)
+      if filter.is_a?(Regexp)
+        element.all_tags.any? { |tag| tag =~ filter }
+      else
+        element.all_tags.include?(filter)
+      end
+    end
+
+    def matching_path?(element, filters)
+      filters.any? do |filtered_path|
+        if filtered_path.is_a?(Regexp)
+          element.get_ancestor(:feature_file).path =~ filtered_path
+        else
+          element.get_ancestor(:feature_file).path == filtered_path
+        end
+      end
+    end
+  end
+end
